[Refactor] #935 Renamed player-personalities-types.h to player-personality-types.h
