+++ Hengband/Hengband/Hengband.vcxproj
@@ -1460,7 +1460,7 @@
     <ClInclude Include="..\..\src\mspell\mspell-type.h" />
     <ClInclude Include="..\..\src\mspell\mspell-util.h" />
     <ClInclude Include="..\..\src\mspell\mspell-ball.h" />
-    <ClInclude Include="..\..\src\player\player-personalities-types.h" />
+    <ClInclude Include="..\..\src\player\player-personality-types.h" />
     <ClInclude Include="..\..\src\player\player-race-types.h" />
     <ClInclude Include="..\..\src\player\player-classes-types.h" />
     <ClInclude Include="..\..\src\spell-kind\spells-detection.h" />
