+++ src/birth/inventory-initializer.cpp
@@ -15,7 +15,7 @@
 #include "object/object-kind-hook.h"
 #include "object/object-kind.h"
 #include "perception/object-perception.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-race-types.h"
 #include "realm/realm-types.h"
 #include "sv-definition/sv-bow-types.h"
