+++ src/birth/quick-start.h
@@ -1,7 +1,7 @@
 ﻿#pragma once
 
 #include "player/player-classes-types.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-race-types.h"
 #include "system/angband.h"
 #include "system/system-variables.h"
