+++ src/cmd-io/cmd-dump.cpp
@@ -23,7 +23,7 @@
 #include "locale/english.h"
 #include "main/sound-of-music.h"
 #include "mutation/mutation-flag-types.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-status-flags.h"
 #include "system/angband-version.h"
 #include "system/floor-type-definition.h"
