+++ src/core/game-play.cpp
@@ -62,7 +62,7 @@
 #include "monster-race/race-indice-types.h"
 #include "monster/monster-util.h"
 #include "player/player-class.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-race-types.h"
 #include "player/player-skill.h"
 #include "player/process-name.h"
