+++ src/mspell/mspell-floor.cpp
@@ -25,7 +25,7 @@
 #include "mspell/mspell-status.h"
 #include "mspell/mspell-util.h"
 #include "mspell/mspell.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-status-flags.h"
 #include "spell-kind/spells-lite.h"
 #include "spell-kind/spells-neighbor.h"
