+++ src/player/player-personality-types.h
