+++ src/player/player-status.cpp
@@ -72,7 +72,7 @@
 #include "player/player-class.h"
 #include "player/player-damage.h"
 #include "player/player-move.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "player/player-personality.h"
 #include "player/player-race-types.h"
 #include "player/player-skill.h"
