+++ src/system/player-type-definition.h
@@ -3,7 +3,7 @@
 #include "player-ability/player-abillity-types.h"
 #include "player/player-race-types.h"
 #include "player/player-classes-types.h"
-#include "player/player-personalities-types.h"
+#include "player/player-personality-types.h"
 #include "system/angband.h"
 #include "system/system-variables.h"
 #include "util/flag-group.h"
